+++ libs/platform/src/lib/value-help-dialog/value-help-dialog/value-help-dialog.component.ts
@@ -6,7 +6,6 @@ import {
     ElementRef,
     EventEmitter,
     Input,
-    isDevMode,
     OnChanges,
     OnDestroy,
     Optional,
@@ -159,11 +158,6 @@ export class PlatformValueHelpDialogComponent<T = any> implements OnChanges, OnD
     /** Loading state */
     @Input()
     loading: boolean | undefined;
-
-    /** @deprecated use `contentDensity` instead */
-    /** The content density for which to render table. 'cozy' | 'compact' | 'condensed' */
-    @Input()
-    searchTableDensity: ContentDensity = ContentDensityEnum.COMPACT;
 
     /** The content density for which to render value help dialog */
     @Input()
@@ -346,15 +340,8 @@ export class PlatformValueHelpDialogComponent<T = any> implements OnChanges, OnD
             }
         }
 
-        if (changes.searchTableDensity || changes.contentDensity) {
+        if (changes.contentDensity) {
             this._contentDensityManuallySet = true;
-
-            if (changes.searchTableDensity) {
-                this.contentDensity = this.searchTableDensity;
-                if (isDevMode()) {
-                    console.warn('"searchTableDensity" is deprecated. Use "contentDensity" instead');
-                }
-            }
         }
     }